refactor server/main and add --merciful flag to avoid the force quit instruments watcher
