+++ lib/server/parser.js
@@ -149,6 +149,16 @@ var args = [
     , nargs: 0
   }],
 
+  [['--merciful', '-m'], {
+    defaultValue: false
+    , dest: 'merciful'
+    , action: 'storeTrue'
+    , required: false
+    , help: "Don't run the watcher process that will force-kill an " +
+            "unresponsive instruments"
+    , nargs: 0
+  }],
+
   [['--app-pkg'], {
     dest: 'androidPackage'
     , defaultValue: null
